code review fixes (#778)
* ABC-01: fix correct metric key

* TXC-01: add missing tx commands

* TXR-01: set correct proposal payload

* move gov tx commands to different file

* update README

* PRI-01: do not set to privileged if already is
